pyproject.toml
[tool.poetry]
name = "boot-python"
version = "0.1.1-alpha"
description = "Prompt-provider plugin for boot-code (Python)."
authors = ["Bordumb <bordumbb@gmail.com>"]
readme = "README.md"
packages = [
  { include = "boot_python", from = "src" }
]

# Include runtime assets (prompts) and optionally proto (dev)
include = [
  { path = "prompts/**/*", format = "sdist" },
  { path = "prompts/**/*", format = "wheel" },
  # proto is not needed at runtime; include only if you want it in sdists:
  { path = "proto/plugin.proto", format = "sdist" }
]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.74.0"
protobuf = "==6.31.1"
tomli = { version = "^2.0.1", python = "<3.11" }  # for TOML parsing on Py<=3.10

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.74.0"
pytest = "^8.0.0"

[tool.poetry.scripts]
boot-python = "boot_python.main:main"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]